- introduced nearby members - every member gets called for rekeying now - avoid exception in CleaningAgent

+++ src/main/java/com/hoccer/talk/server/cleaning/CleaningAgent.java
@@ -252,7 +252,7 @@ public class CleaningAgent {
 
         // delete attached file if there is one
         String fileId = message.getAttachmentFileId();
-        if (fileId != null) {
+        if (fileId != null && mFilecache != null) {   // TODO: mFilecache is sometimes null for unknown reasons, should be fixed
             mFilecache.deleteFile(fileId);
         }
